+++ R/convertLayer.R
@@ -11,13 +11,26 @@
 #' If shapefiles have no projection information they are assumed to 'geographic' coordinates  and will be assigned
 #' \code{"+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"} before being returned.
 #' @return An object of class \code{SpatialPolygonsDataFrame}
-convertLayer <- function(dsn="", layerName="", encoding=NULL) {
-
+convertLayer <- function(
+  dsn = "", 
+  layerName = "", 
+  encoding = NULL
+) {
+  
   # readOGR does not interpret '~' so do that with path.expand()
   dsn <- path.expand(dsn)
+  
+  # NOTE:  Chosoe 'use_iconv' setting based on encoding (see ?rgdal::readOGR)
 
-  data_projected <- rgdal::readOGR(dsn=dsn, layer=layerName, stringsAsFactors=FALSE, encoding=encoding) 
-
+  # Read in projected data
+  data_projected <- rgdal::readOGR(
+    dsn = dsn, 
+    layer = layerName, 
+    stringsAsFactors = FALSE, 
+    encoding = encoding,
+    use_iconv = ifelse(is.null(encoding), FALSE, TRUE)
+  ) 
+  
   # Assign or reproject to standard projection
   if ( is.na(sp::proj4string(data_projected)) ) {
     # NOTE:  shapefiles used in convertHMSSmoke() have no projection whatsoever
@@ -26,7 +39,7 @@ convertLayer <- function(dsn="", layerName="", encoding=NULL) {
   } else {
     SPDF <- sp::spTransform(data_projected, sp::CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"))
   }
-
+  
   return(SPDF)
 }